Calculate rowStatusColumn from column definition; bug fix re: overwrite
There was an optimization where the row status column number was saved
in the provider, in addition to marking that column as the row status
column. With this change, when the row status column number is needed,
it is ascertained by scanning the column definitions for the one
marked as the row status column. This is a bit slower, but is
infrequently called.

Also fixed here is the case where a SET is issued with "createAndGo",
which results in the row being created and the status column being set
to "active"... followed by a SET with "createAndWait" which should
have returned an error, but instead changed the status value to
"notInService"