Disable erroneous use of auto-padding in DES decrypt
SNMP does not use the padding scheme that nodejs uses when its
autoPadding option is enabled, and leaving this option enabled when the
sender uses zero-padding (which is typical) will either throw an
exception or result in a corrupted packet.

NOTE: leaving auto-padding enabled in encrypt is not harmful (since SNMP
does not care about the padding bytes), although it results in a packet
which is 8 bytes longer than needed if the original contextPdu was
already a multiple of 8 bytes.