Mib lexer: Split name followed by group into separate symbols
This supports the valid syntax where a type constraint specified in
parens or braces joins the type specifier with no whitespace in between.
E.g., previously
   DisplayString(SIZE (2..10))
was parsed as a single symbol.  There was a special case for INTEGER,
but this did not extend to Integer32 etc.  This patch generalizes the
INTEGER case to tokenize upon seeing a '(' or '{' that open an outer
group.