ModuleStore: Allow to configure base modules.
This provides a way to workaround the change of behaviour introduced by
the deferred registration update.  Constraints for SNMPv2-TC defined
textual conventions like DisplayString seem to get preempted by the
subsequent definitions as plain OCTET STRING in RFC MIBs.

This feature allows the user to select upon which base MIBs they depend
explicitly so as not to get any surprises or unexpected type overrides.

Example usage to select just the SNMPv2 MIBs:
   snmp.createModuleStore({
     baseModules: [
       'SNMPv2-SMI',
       'SNMPv2-CONF',
       'SNMPv2-TC',
       'SNMPv2-MIB',
     ],
   });